+++ src/sphinx_codeautolink/extension/resolve.py
@@ -1,72 +1,126 @@
 """Resolve import locations and type hints."""
+from dataclasses import dataclass
 from functools import lru_cache
 from importlib import import_module
-from typing import Optional, Tuple, Any, Union, Callable
+from inspect import isclass, isroutine
+from typing import Optional, Tuple, Any, Union, Callable, List
 from ..parse import Name, NameBreak
 
 
-def resolve_location(chain: Name, inventory) -> Optional[str]:
+def resolve_location(chain: Name, inventory) -> str:
     """Find the final type that a name refers to."""
     comps = []
+    cursor = None
     for comp in chain.import_components:
-        if comp == NameBreak.call:
-            new = locate_type(tuple(comps), inventory, ends_with_call=True)
-            if new is None:
-                return
-            comps = new.split('.')
-        else:
+        if comp != NameBreak.call:
             comps.append(comp)
+            continue
 
-    imported_loc = locate_type(tuple(comps), inventory, ends_with_call=False)
-    return imported_loc or '.'.join(comps)
+        if cursor is None:
+            comps, cursor = make_cursor(comps)
+
+        cursor = locate_type(cursor, tuple(comps), inventory)
+        if cursor is None:
+            raise CouldNotResolve()
+
+        call_value(cursor)
+        comps = []
+
+    if cursor is None:
+        try:
+            comps, cursor = make_cursor(comps)
+        except CouldNotResolve:
+            if '.'.join(comps) in inventory:
+                # Last ditch effort to locate based on the string only
+                return '.'.join(comps)
+            raise
+
+    cursor = locate_type(cursor, tuple(comps), inventory)
+    return cursor.location if cursor is not None else None
+
+
+class CouldNotResolve(Exception):
+    """Could not resolve type to inventory."""
+
+
+@dataclass
+class Cursor:
+    """Cursor to follow imports, attributes and calls to the final type."""
+
+    location: str
+    value: Any
+    instance: bool
+
+
+def make_cursor(components: List[str]) -> Tuple[List[str], Cursor]:
+    """Divide components into module and rest, create cursor for following the rest."""
+    value, index = closest_module(tuple(components))
+    if value is None or index is None:
+        raise CouldNotResolve()
+    location = '.'.join(components[:index])
+    return components[index:], Cursor(location, value, False)
 
 
 def locate_type(
-    components: Tuple[str], inventory, ends_with_call: bool
-) -> Optional[str]:
+    cursor: Cursor, components: Tuple[str, ...], inventory
+) -> Cursor:
     """Find type hint and resolve to new location."""
-    value, index = closest_module(components)
-    if index is None or index == len(components):
-        return
-    remaining = components[index:]
-    location = '.'.join(components[:index])
+    previous = cursor
+    for i, component in enumerate(components):
+        cursor = Cursor(
+            cursor.location + '.' + component,
+            getattr(cursor.value, component, None),
+            cursor.instance,
+        )
 
-    prev_val = None
-    for component in remaining:
-        value = getattr(value, component, None)
-        location += '.' + component
-        if value is None:
-            return
+        if cursor.value is None:
+            raise CouldNotResolve()
 
-        if isinstance(value, type) or (callable(value) and location not in inventory):
+        if isclass(cursor.value):
+            cursor.instance = False
+
+        if (
+            isclass(cursor.value)
+            or (isroutine(cursor.value) and cursor.location not in inventory)
+        ):
+            # Normalise location of type or imported function
             try:
-                location = fully_qualified_name(value)
+                cursor.location = fully_qualified_name(cursor.value)
             except (AttributeError, TypeError):
                 # Odd construct encountered: don't try to be clever but continue
                 pass
 
-        if isinstance(prev_val, type) and location not in inventory:
-            for val in prev_val.mro():
-                loc = fully_qualified_name(val) + '.' + component
-                if loc in inventory:
-                    return locate_type(tuple(loc.split('.')), inventory, ends_with_call)
+        if isclass(previous.value) and cursor.location not in inventory:
+            for val in previous.value.mro():
+                name = fully_qualified_name(val)
+                if name + '.' + component in inventory:
+                    previous.location = name
+                    return locate_type(previous, components[i:], inventory)
 
-        prev_val = value
+        previous = cursor
 
-    # A possible function / method call needs to be last in the chain.
-    # Otherwise we might follow return types on function attribute access.
-    if (
-        ends_with_call
-        and remaining
-        and callable(value)
-        and not isinstance(value, type)
-    ):
-        location = follow_return_annotation(value)
-
-    return location
+    return cursor
 
 
-def follow_return_annotation(func: Callable) -> Optional[str]:
+def call_value(cursor: Cursor):
+    """Call class, instance or function."""
+    if isclass(cursor.value) and not cursor.instance:
+        # class definition: "instantiate" class
+        cursor.instance = True
+        return
+
+    if callable(cursor.value) and not isroutine(cursor.value):
+        # callable class instance
+        cursor.value = cursor.value.__call__
+    elif not isroutine(cursor.value):
+        raise CouldNotResolve()  # not a function either
+
+    cursor.value = get_return_annotation(cursor.value)
+    cursor.location = fully_qualified_name(cursor.value)
+    cursor.instance = True
+
+
+def get_return_annotation(func: Callable) -> Optional[type]:
     """Determine the target of a function return type hint."""
     annotations = getattr(func, '__annotations__', {})
     ret_annotation = annotations.get('return', None)
@@ -92,9 +146,9 @@ def follow_return_annotation(func: Callable) -> Optional[str]:
         or not isinstance(ret_annotation, type)
         or hasattr(ret_annotation, '__origin__')
     ):
-        return
+        raise CouldNotResolve()
 
-    return fully_qualified_name(ret_annotation)
+    return ret_annotation
 
 
 def fully_qualified_name(thing: Union[type, Callable]) -> str:
@@ -103,15 +157,18 @@ def fully_qualified_name(thing: Union[type, Callable]) -> str:
 
 
 @lru_cache(maxsize=None)
-def closest_module(components: Tuple[str]) -> Tuple[Any, Optional[int]]:
+def closest_module(components: Tuple[str, ...]) -> Tuple[Any, Optional[int]]:
     """Find closest importable module."""
-    mod = None
-    for i in range(1, len(components) + 1):
-        try:
-            mod = import_module('.'.join(components[:i]))
-        except ImportError:
-            break
-    else:
+    try:
+        mod = import_module(components[0])
+    except ImportError:
         return None, None
 
-    return mod, i - 1
+    for i in range(1, len(components)):
+        try:
+            mod = import_module('.'.join(components[:i + 1]))
+        except ImportError:
+            # import failed, exclude previously added item
+            return mod, i
+    # imports succeeded, include all items
+    return mod, len(components)